unify all handler function in same file
